animateGarageDoor

+++ src/cmd_interface/resources.rs
@@ -13,7 +13,7 @@ pub struct InfoContainerData {
     pub active: bool,
     pub name: String,
     pub cost: i32,
-    pub build_time: i32,
+    pub build_time: u64,
     pub hp: Option<i32>,
     pub dmg: Option<i32>,
     pub speed: Option<f32>,
